refactor: extract wallet delegation module and spec
